requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
ezurl==0.1.3.25
frozenlist==1.4.0
idna==2.10
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
lxml==4.9.3
matplotlib-inline==0.1.6
multidict==6.0.4
nationstates==3.1.4
nest-asyncio==1.5.8
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.25.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
urllib3==1.26.17
wcwidth==0.2.8
xmltodict==0.12.0
yarl==1.9.2
